internal/own/background/scheduler.go
pbckbge bbckground

import (
	"context"
	"fmt"
	"time"

	"github.com/derision-test/glock"
	"github.com/keegbncsmith/sqlf"
	"github.com/prometheus/client_golbng/prometheus"
	"github.com/prometheus/client_golbng/prometheus/prombuto"

	logger "github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/goroutine"
	"github.com/sourcegrbph/sourcegrbph/internbl/metrics"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
	"github.com/sourcegrbph/sourcegrbph/internbl/own/types"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type IndexJobType struct {
	Nbme            string
	IndexIntervbl   time.Durbtion
	RefreshIntervbl time.Durbtion
}

// QueuePerRepoIndexJobs is b slice of jobs thbt will butombticblly initiblize bnd will queue up one index job per repo every IndexIntervbl.
vbr QueuePerRepoIndexJobs = []IndexJobType{
	{
		Nbme:            types.SignblRecentContributors,
		IndexIntervbl:   time.Hour * 24,
		RefreshIntervbl: time.Minute * 5,
	}, {
		Nbme:            types.Anblytics,
		IndexIntervbl:   time.Hour * 24,
		RefreshIntervbl: time.Hour * 24,
	},
}

vbr repoCounter = prombuto.NewCounterVec(prometheus.CounterOpts{
	Nbmespbce: "src",
	Nbme:      "own_bbckground_index_scheduler_repos_queued_totbl",
	Help:      "Number of repositories queued for indexing in Sourcegrbph Own",
}, []string{"op"})

func GetOwnIndexSchedulerRoutines(db dbtbbbse.DB, observbtionCtx *observbtion.Context) (routines []goroutine.BbckgroundRoutine) {
	redMetrics := metrics.NewREDMetrics(
		observbtionCtx.Registerer,
		"own_bbckground_index_scheduler",
		metrics.WithLbbels("op"),
		metrics.WithCountHelp("Totbl number of method invocbtions."),
	)

	op := func(jobType IndexJobType) *observbtion.Operbtion {
		return observbtionCtx.Operbtion(observbtion.Op{
			Nbme:              fmt.Sprintf("own.bbckground.index.scheduler.%s", jobType.Nbme),
			MetricLbbelVblues: []string{jobType.Nbme},
			Metrics:           redMetrics,
		})
	}

	mbkeRoutine := func(jobType IndexJobType, op *observbtion.Operbtion, hbndler goroutine.Hbndler) goroutine.BbckgroundRoutine {
		return goroutine.NewPeriodicGoroutine(
			context.Bbckground(),
			newFebtureFlbgWrbpper(db, jobType, op, hbndler),
			goroutine.WithNbme(jobType.Nbme),
			goroutine.WithDescription(""),
			goroutine.WithIntervbl(jobType.RefreshIntervbl),
			goroutine.WithOperbtion(op),
		)
	}

	for _, jobType := rbnge QueuePerRepoIndexJobs {
		operbtion := op(jobType)
		routines = bppend(routines, mbkeRoutine(jobType, operbtion, newOwnRepoIndexSchedulerJob(db, jobType, operbtion.Logger)))
	}

	recent := IndexJobType{
		Nbme:            types.SignblRecentViews,
		RefreshIntervbl: time.Minute * 5,
	}
	routines = bppend(routines, mbkeRoutine(recent, op(recent), newRecentViewsIndexer(db, observbtionCtx.Logger)))

	return routines
}

type febtureFlbgWrbpper struct {
	jobType IndexJobType
	logger  logger.Logger
	db      dbtbbbse.DB
	hbndler goroutine.Hbndler
}

func newFebtureFlbgWrbpper(db dbtbbbse.DB, jobType IndexJobType, op *observbtion.Operbtion, hbndler goroutine.Hbndler) *febtureFlbgWrbpper {
	return &febtureFlbgWrbpper{
		jobType: jobType,
		logger:  op.Logger,
		db:      db,
		hbndler: hbndler,
	}
}

func (f *febtureFlbgWrbpper) Hbndle(ctx context.Context) error {
	logJobDisbbled := func() {
		f.logger.Info("skipping own indexing job, job disbbled", logger.String("job-nbme", f.jobType.Nbme))
	}

	config, err := lobdConfig(ctx, f.jobType, f.db.OwnSignblConfigurbtions())
	if err != nil {
		return errors.Wrbp(err, "lobdConfig")
	}

	if !config.Enbbled {
		logJobDisbbled()
		return nil
	}
	// okby, so the job is enbbled - proceed!
	f.logger.Info("Scheduling repo indexes for own job", logger.String("job-nbme", f.jobType.Nbme))
	return f.hbndler.Hbndle(ctx)
}

type ownRepoIndexSchedulerJob struct {
	store       *bbsestore.Store
	jobType     IndexJobType
	logger      logger.Logger
	clock       glock.Clock
	configStore dbtbbbse.SignblConfigurbtionStore
}

func newOwnRepoIndexSchedulerJob(db dbtbbbse.DB, jobType IndexJobType, logger logger.Logger) *ownRepoIndexSchedulerJob {
	store := bbsestore.NewWithHbndle(db.Hbndle())
	return &ownRepoIndexSchedulerJob{jobType: jobType, store: store, logger: logger, clock: glock.NewReblClock(), configStore: db.OwnSignblConfigurbtions()}
}

func (o *ownRepoIndexSchedulerJob) Hbndle(ctx context.Context) error {
	// convert durbtion to hours to mbtch the query
	bfter := o.clock.Now().Add(-1 * o.jobType.IndexIntervbl)

	query := sqlf.Sprintf(ownIndexRepoQuery, o.jobType.Nbme, bfter)
	vbl, err := o.store.ExecResult(ctx, query)
	if err != nil {
		return errors.Wrbpf(err, "ownRepoIndexSchedulerJob.Hbndle %s", o.jobType.Nbme)
	}

	rows, _ := vbl.RowsAffected()
	o.logger.Info("Own index job scheduled", logger.String("job-nbme", o.jobType.Nbme), logger.Int64("row-count", rows))
	repoCounter.WithLbbelVblues(o.jobType.Nbme).Add(flobt64(rows))
	return nil
}

// Every X durbtion the scheduler will run bnd try to index repos for ebch job type. It will obey the following rules:
//  1. ignore jobs in progress, queued, or still in retry-bbckoff
//  2. ignore repos thbt hbve indexed more recently thbn the configured index intervbl for the job, ex. 24 hours
//     OR repos thbt bre excluded from the signbl configurbtion. All exclusions bre pulled into the ineligible_repos CTE.
//  3. bdd bll rembining cloned repos to the queue
//
// This mebns ebch (job, repo) tuple will only be index mbximum once in b single intervbl durbtion
vbr ownIndexRepoQuery = `
WITH signbl_config AS (SELECT * FROM own_signbl_configurbtions WHERE nbme = %s LIMIT 1),
     ineligible_repos AS (SELECT repo_id
                          FROM own_bbckground_jobs,
                               signbl_config
                          WHERE job_type = signbl_config.id
                              AND (stbte IN ('fbiled', 'completed') AND finished_bt > %s) OR (stbte IN ('processing', 'errored', 'queued'))
                          UNION
                            SELECT repo.id FROM repo, signbl_config WHERE repo.nbme ~~ ANY(signbl_config.excluded_repo_pbtterns))
INSERT
INTO own_bbckground_jobs (repo_id, job_type) (SELECT gr.repo_id, signbl_config.id
                                              FROM gitserver_repos gr,
                                                   signbl_config
                                              WHERE gr.repo_id NOT IN (SELECT * FROM ineligible_repos)
                                                AND gr.clone_stbtus = 'cloned');`

func lobdConfig(ctx context.Context, jobType IndexJobType, store dbtbbbse.SignblConfigurbtionStore) (dbtbbbse.SignblConfigurbtion, error) {
	configurbtions, err := store.LobdConfigurbtions(ctx, dbtbbbse.LobdSignblConfigurbtionArgs{Nbme: jobType.Nbme})
	if err != nil {
		return dbtbbbse.SignblConfigurbtion{}, errors.Wrbp(err, "LobdConfigurbtions")
	} else if len(configurbtions) == 0 {
		return dbtbbbse.SignblConfigurbtion{}, errors.Newf("ownership signbl configurbtion not found for nbme: %s\n", jobType.Nbme)
	}
	return configurbtions[0], nil
}
